@@ -1,11 +1,11 @@
 {
   "name": "wonrial",
-  "version": "25.06.27",
+  "version": "25.06.29",
   "description": "WONRIAL ENTERPRISES LTD",
   "author": "yurapa",
-  "private": false,
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
@@ -14,39 +14,44 @@
     "code:format": "prettier './**/**.{ts,tsx,css,scss}' --write"
   },
   "dependencies": {
-    "@ai-sdk/openai": "^0.0.63",
-    "@next/third-parties": "^14.2.14",
-    "@vercel/speed-insights": "^1.1.0",
-    "ai": "^3.4.7",
+    "@ai-sdk/openai": "^1.3.22",
+    "@next/eslint-plugin-next": "^15.3.4",
+    "@next/third-parties": "^15.3.4",
+    "@vercel/speed-insights": "^1.2.0",
+    "ai": "^4.3.16",
     "classnames": "^2.5.1",
-    "i18next": "^23.7.6",
-    "i18next-browser-languagedetector": "^7.2.0",
+    "i18next": "^25.2.1",
+    "i18next-browser-languagedetector": "^8.2.0",
     "i18next-resources-to-backend": "^1.2.0",
     "lodash": "^4.17.21",
-    "next": "^14.2.14",
-    "next-themes": "^0.3.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-i18next": "^13.4.1",
-    "react-icons": "^5.3.0"
+    "next": "^15.3.4",
+    "next-themes": "^0.4.6",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-i18next": "^15.5.3",
+    "react-icons": "^5.5.0"
   },
   "devDependencies": {
-    "@types/lodash": "^4.17.14",
-    "@types/node": "^22.7.4",
-    "@types/react": "^18.3.10",
-    "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^6.11.0",
-    "@typescript-eslint/parser": "^6.11.0",
-    "autoprefixer": "^10.4.20",
-    "eslint": "^8.53.0",
-    "eslint-config-next": "14.0.2",
-    "eslint-config-prettier": "^9.0.0",
-    "postcss": "^8.4.47",
-    "prettier": "^3.3.3",
-    "prettier-plugin-tailwindcss": "^0.6.8",
-    "sass": "^1.79.4",
-    "tailwindcss": "^3.4.13",
-    "typescript": "^5.6.2"
+    "@eslint/js": "^9.30.0",
+    "@tailwindcss/postcss": "^4.1.11",
+    "@types/lodash": "^4.17.19",
+    "@types/node": "^24.0.6",
+    "@types/react": "^19.0.0",
+    "@types/react-dom": "^19.0.0",
+    "@typescript-eslint/eslint-plugin": "^8.35.0",
+    "@typescript-eslint/parser": "^8.35.0",
+    "autoprefixer": "^10.4.21",
+    "eslint": "^9.30.0",
+    "eslint-config-next": "15.3.4",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "postcss": "^8.5.6",
+    "prettier": "^3.6.2",
+    "prettier-plugin-tailwindcss": "^0.6.13",
+    "sass": "^1.89.2",
+    "tailwindcss": "^4.1.11",
+    "typescript": "^5.8.3"
   },
   "files": [
     ".browserslistrc",
