@@ -1,6 +1,8 @@
-module.exports = {
+const config = {
   plugins: {
-    tailwindcss: {},
+    '@tailwindcss/postcss': {},
     autoprefixer: {},
   },
 };
+
+export default config;
