@@ -10,7 +10,8 @@ export const metadata: Metadata = {
   title: 'Contact Us :: Wonrial',
 };
 
-export default async function Contact({ params: { locale } }: { params: { locale: string } }) {
+export default async function Contact({ params }: { params: Promise<{ locale: string }> }) {
+  const { locale } = await params;
   const { t } = await createTranslation(locale, 'common');
 
   return (
@@ -20,8 +21,8 @@ export default async function Contact({ params: { locale } }: { params: { locale
           <div className="-mx-4 flex flex-wrap items-center">
             <div className="w-full px-4 md:w-8/12 lg:w-7/12">
               <div className="mb-8 max-w-[570px] md:mb-0 lg:mb-12">
-                <h1 className="mb-5 text-2xl font-bold text-black dark:text-white sm:text-3xl">{t('contact.title')}</h1>
-                <p className="text-base font-medium leading-relaxed text-body-color">Our Contacts,</p>
+                <h1 className="mb-5 text-2xl font-bold text-black sm:text-3xl dark:text-white">{t('contact.title')}</h1>
+                <p className="text-body-color text-base leading-relaxed font-medium">Our Contacts,</p>
               </div>
             </div>
           </div>
@@ -33,57 +34,57 @@ export default async function Contact({ params: { locale } }: { params: { locale
           <div className="-mx-4 flex flex-wrap">
             <div className="w-full px-4 lg:w-7/12 xl:w-8/12">
               <div
-                className="wow fadeInUp mb-12 rounded-sm bg-white px-8 py-11 shadow-three dark:bg-gray-dark sm:p-[55px] lg:mb-5 lg:px-8 xl:p-[55px]"
+                className="wow fadeInUp shadow-three dark:bg-gray-dark mb-12 rounded-sm bg-white px-8 py-11 sm:p-[55px] lg:mb-5 lg:px-8 xl:p-[55px]"
                 data-wow-delay=".15s
               "
               >
-                <h2 className="mb-3 text-2xl font-bold text-black dark:text-white sm:text-3xl lg:text-2xl xl:text-3xl">
+                <h2 className="mb-3 text-2xl font-bold text-black sm:text-3xl lg:text-2xl xl:text-3xl dark:text-white">
                   Need Help? Open a Ticket
                 </h2>
-                <p className="mb-12 text-base font-medium text-body-color">
+                <p className="text-body-color mb-12 text-base font-medium">
                   Our support team will get back to you ASAP via email.
                 </p>
                 <form>
                   <div className="-mx-4 flex flex-wrap">
                     <div className="w-full px-4 md:w-1/2">
                       <div className="mb-8">
-                        <label htmlFor="name" className="mb-3 block text-sm font-medium text-dark dark:text-white">
+                        <label htmlFor="name" className="text-dark mb-3 block text-sm font-medium dark:text-white">
                           Your Name
                         </label>
                         <input
                           type="text"
                           placeholder="Enter your name"
-                          className="w-full rounded-sm border border-stroke bg-[#f8f8f8] px-6 py-3 text-base text-body-color outline-none focus:border-primary dark:border-transparent dark:bg-[#2C303B] dark:text-body-color-dark dark:shadow-two dark:focus:border-primary dark:focus:shadow-none"
+                          className="border-stroke text-body-color focus:border-primary dark:text-body-color-dark dark:shadow-two dark:focus:border-primary w-full rounded-sm border bg-[#f8f8f8] px-6 py-3 text-base outline-none dark:border-transparent dark:bg-[#2C303B] dark:focus:shadow-none"
                         />
                       </div>
                     </div>
                     <div className="w-full px-4 md:w-1/2">
                       <div className="mb-8">
-                        <label htmlFor="email" className="mb-3 block text-sm font-medium text-dark dark:text-white">
+                        <label htmlFor="email" className="text-dark mb-3 block text-sm font-medium dark:text-white">
                           Your Email
                         </label>
                         <input
                           type="email"
                           placeholder="Enter your email"
-                          className="w-full rounded-sm border border-stroke bg-[#f8f8f8] px-6 py-3 text-base text-body-color outline-none focus:border-primary dark:border-transparent dark:bg-[#2C303B] dark:text-body-color-dark dark:shadow-two dark:focus:border-primary dark:focus:shadow-none"
+                          className="border-stroke text-body-color focus:border-primary dark:text-body-color-dark dark:shadow-two dark:focus:border-primary w-full rounded-sm border bg-[#f8f8f8] px-6 py-3 text-base outline-none dark:border-transparent dark:bg-[#2C303B] dark:focus:shadow-none"
                         />
                       </div>
                     </div>
                     <div className="w-full px-4">
                       <div className="mb-8">
-                        <label htmlFor="message" className="mb-3 block text-sm font-medium text-dark dark:text-white">
+                        <label htmlFor="message" className="text-dark mb-3 block text-sm font-medium dark:text-white">
                           Your Message
                         </label>
                         <textarea
                           name="message"
                           rows={5}
                           placeholder="Enter your Message"
-                          className="w-full resize-none rounded-sm border border-stroke bg-[#f8f8f8] px-6 py-3 text-base text-body-color outline-none focus:border-primary dark:border-transparent dark:bg-[#2C303B] dark:text-body-color-dark dark:shadow-two dark:focus:border-primary dark:focus:shadow-none"
+                          className="border-stroke text-body-color focus:border-primary dark:text-body-color-dark dark:shadow-two dark:focus:border-primary w-full resize-none rounded-sm border bg-[#f8f8f8] px-6 py-3 text-base outline-none dark:border-transparent dark:bg-[#2C303B] dark:focus:shadow-none"
                         ></textarea>
                       </div>
                     </div>
                     <div className="w-full px-4">
-                      <button className="rounded-sm bg-primary px-9 py-4 text-base font-medium text-white shadow-submit duration-300 hover:bg-primary/90 dark:shadow-submit-dark">
+                      <button className="bg-primary shadow-submit hover:bg-primary/90 dark:shadow-submit-dark rounded-sm px-9 py-4 text-base font-medium text-white duration-300">
                         Submit Ticket
                       </button>
                     </div>
@@ -105,6 +106,7 @@ export default async function Contact({ params: { locale } }: { params: { locale
           width={100}
           height={100}
           sizes="100vw"
+          priority
           style={{
             width: '100%',
             height: 'auto',
