@@ -11,7 +11,8 @@ export const metadata: Metadata = {
   title: 'Our Services :: Wonrial',
 };
 
-export default async function Services({ params: { locale } }: { params: { locale: string } }) {
+export default async function Services({ params }: { params: Promise<{ locale: string }> }) {
+  const { locale } = await params;
   const { t } = await createTranslation(locale, 'common');
 
   return (
@@ -22,10 +23,10 @@ export default async function Services({ params: { locale } }: { params: { local
             <div className="-mx-4 flex flex-wrap items-center">
               <div className="w-full px-4 md:w-8/12 lg:w-7/12">
                 <div className="mb-8 max-w-[570px] md:mb-0 lg:mb-12">
-                  <h1 className="mb-5 text-2xl font-bold text-black dark:text-white sm:text-3xl">
+                  <h1 className="mb-5 text-2xl font-bold text-black sm:text-3xl dark:text-white">
                     {t('services.title')}
                   </h1>
-                  <p className="text-base font-medium leading-relaxed text-body-color">Our services,</p>
+                  <p className="text-body-color text-base leading-relaxed font-medium">Our services,</p>
                 </div>
               </div>
             </div>
