@@ -12,17 +12,17 @@ export function AIChat() {
         <div className="-mx-4 flex flex-wrap">
           <div className="w-full px-4">
             <div
-              className="wow fadeInUp mb-12 rounded-sm bg-white px-8 py-11 shadow-three dark:bg-gray-dark sm:p-[55px] lg:mb-5 lg:px-8 xl:p-[55px]"
+              className="wow fadeInUp shadow-three dark:bg-gray-dark mb-12 rounded-sm bg-white px-8 py-11 sm:p-[55px] lg:mb-5 lg:px-8 xl:p-[55px]"
               data-wow-delay=".15s
               "
             >
               {isLoading && (
-                <h2 className="mb-3 text-2xl font-bold text-black dark:text-white sm:text-3xl lg:text-2xl xl:text-3xl">
+                <h2 className="mb-3 text-2xl font-bold text-black sm:text-3xl lg:text-2xl xl:text-3xl dark:text-white">
                   Loading...
                 </h2>
               )}
 
-              <div className="mb-12 text-base font-medium text-body-color dark:text-white">
+              <div className="text-body-color mb-12 text-base font-medium dark:text-white">
                 {messages.map((m) => (
                   <div key={m.id} className="mb-6 whitespace-pre-wrap">
                     {m.role === 'user' ? 'User: ' : 'AI: '}
@@ -45,7 +45,7 @@ export function AIChat() {
                         handleSubmit(e);
                       }
                     }}
-                    className="mr-4 w-full resize-none rounded-sm border border-stroke bg-[#f8f8f8] px-6 py-3 text-base text-body-color outline-none focus:border-primary dark:border-transparent dark:bg-[#2C303B] dark:text-body-color-dark dark:shadow-two dark:focus:border-primary dark:focus:shadow-none"
+                    className="border-stroke text-body-color focus:border-primary dark:text-body-color-dark dark:shadow-two dark:focus:border-primary mr-4 w-full resize-none rounded-sm border bg-[#f8f8f8] px-6 py-3 text-base outline-none dark:border-transparent dark:bg-[#2C303B] dark:focus:shadow-none"
                   />
                   <AiChatButton />
                 </div>
