@@ -9,19 +9,19 @@ export default async function Hero({ locale }: { locale: LocaleTypes }) {
   return (
     <section
       id="home"
-      className="relative z-10 overflow-hidden bg-white pb-16 pt-[120px] dark:bg-gray-dark md:pb-[120px] md:pt-[150px] xl:pb-[160px] xl:pt-[180px] 2xl:pb-[200px] 2xl:pt-[210px]"
+      className="dark:bg-gray-dark relative z-10 overflow-hidden bg-white pt-[120px] pb-16 md:pt-[150px] md:pb-[120px] xl:pt-[180px] xl:pb-[160px] 2xl:pt-[210px] 2xl:pb-[200px]"
     >
       <div className="container">
         <div className="-mx-4 flex flex-wrap">
           <div className="w-full px-4">
             <div className="wow fadeInUp mx-auto max-w-[800px] text-center" data-wow-delay=".2s">
-              <h1 className="mb-5 text-3xl font-bold leading-tight text-black dark:text-white sm:text-4xl sm:leading-tight md:text-5xl md:leading-tight">
+              <h1 className="mb-5 text-3xl leading-tight font-bold text-black sm:text-4xl sm:leading-tight md:text-5xl md:leading-tight dark:text-white">
                 {t('homePageBanner.title')}
               </h1>
-              <p className="mb-12 text-base !leading-relaxed text-body-color dark:text-body-color-dark sm:text-lg md:text-xl">
+              <p className="text-body-color dark:text-body-color-dark mb-12 text-base !leading-relaxed sm:text-lg md:text-xl">
                 {t('homePageBanner.subTitle')}
               </p>
-              <div className="flex flex-col items-center justify-center space-y-4 sm:flex-row sm:space-x-4 sm:space-y-0">
+              <div className="flex flex-col items-center justify-center space-y-4 sm:flex-row sm:space-y-0 sm:space-x-4">
                 <EventButton />
                 <Link
                   href="/"
@@ -34,7 +34,7 @@ export default async function Hero({ locale }: { locale: LocaleTypes }) {
           </div>
         </div>
       </div>
-      <div className="absolute right-0 top-0 z-[-1] opacity-30 lg:opacity-100">
+      <div className="absolute top-0 right-0 z-[-1] opacity-30 lg:opacity-100">
         <svg width="450" height="556" viewBox="0 0 450 556" fill="none" xmlns="http://www.w3.org/2000/svg">
           <circle cx="277" cy="63" r="225" fill="url(#paint0_linear_25:217)" />
           <circle cx="17.9997" cy="182" r="18" fill="url(#paint1_radial_25:217)" />
