@@ -13,10 +13,16 @@ const crossIcon = (
 const OfferList = ({ text, status }: { text: string; status: 'active' | 'inactive' }) => {
   return (
     <div className="mb-3 flex items-center">
-      <span className="mr-3 flex h-[18px] w-full max-w-[18px] items-center justify-center rounded-full bg-primary bg-opacity-10 text-primary">
+      <span
+        className={`mr-3 flex h-[18px] w-full max-w-[18px] items-center justify-center rounded-full ${
+          status === 'active'
+            ? 'bg-primary/10 text-primary [&>svg]:fill-primary'
+            : 'bg-red-500/10 text-red-500 [&>svg]:fill-red-500'
+        }`}
+      >
         {status === 'active' ? checkIcon : crossIcon}
       </span>
-      <p className="m-0 text-base font-medium text-body-color">{text}</p>
+      <p className="text-body-color m-0 text-base font-medium">{text}</p>
     </div>
   );
 };
