@@ -22,7 +22,7 @@ const Pricing = () => {
             <span
               onClick={() => setIsMonthly(true)}
               className={`${
-                isMonthly ? 'pointer-events-none text-primary' : 'text-dark dark:text-white'
+                isMonthly ? 'text-primary pointer-events-none' : 'text-dark dark:text-white'
               } mr-4 cursor-pointer text-base font-semibold`}
             >
               Monthly
@@ -33,7 +33,7 @@ const Pricing = () => {
                 <div
                   className={`${
                     isMonthly ? '' : 'translate-x-full'
-                  } shadow-switch-1 absolute left-0 top-[-4px] flex h-7 w-7 items-center justify-center rounded-full bg-primary transition`}
+                  } shadow-switch-1 bg-primary absolute top-[-4px] left-0 flex h-7 w-7 items-center justify-center rounded-full transition`}
                 >
                   <span className="active h-4 w-4 rounded-full bg-white"></span>
                 </div>
@@ -42,7 +42,7 @@ const Pricing = () => {
             <span
               onClick={() => setIsMonthly(false)}
               className={`${
-                isMonthly ? 'text-dark dark:text-white' : 'pointer-events-none text-primary'
+                isMonthly ? 'text-dark dark:text-white' : 'text-primary pointer-events-none'
               } ml-4 cursor-pointer text-base font-semibold`}
             >
               Yearly
