@@ -18,14 +18,14 @@ const ReadyToHelp = () => {
           <div className="w-full px-4">
             <div className="wow fadeInUp mx-auto max-w-[770px] overflow-hidden rounded-md" data-wow-delay=".15s">
               <div className="relative aspect-[77/40] items-center justify-center">
-                <Image src="/images/ready-to-help.jpg" alt="video image" fill />
+                <Image src="/images/ready-to-help.jpg" alt="video image" fill sizes="(max-width: 768px) 100vw, 770px" />
               </div>
             </div>
           </div>
         </div>
       </div>
 
-      <div className="absolute bottom-0 left-0 right-0 z-[-1] h-full w-full bg-[url(/images/shape.svg)] bg-cover bg-center bg-no-repeat"></div>
+      <div className="absolute right-0 bottom-0 left-0 z-[-1] h-full w-full bg-[url(/images/shape.svg)] bg-cover bg-center bg-no-repeat"></div>
     </section>
   );
 };
