@@ -18,10 +18,10 @@ const SectionTitle = ({
         data-wow-delay=".1s"
         style={{ maxWidth: width, marginBottom: mb }}
       >
-        <h2 className="mb-4 text-3xl font-bold !leading-tight text-black dark:text-white sm:text-4xl md:text-[45px]">
+        <h2 className="mb-4 text-3xl !leading-tight font-bold text-black sm:text-4xl md:text-[45px] dark:text-white">
           {title}
         </h2>
-        <p className="text-base !leading-relaxed text-body-color md:text-lg">{paragraph}</p>
+        <p className="text-body-color text-base !leading-relaxed md:text-lg">{paragraph}</p>
       </div>
     </>
   );
