@@ -22,20 +22,20 @@ const Testimonial = ({ testimonial }: { testimonial: TestimonialItem }) => {
   return (
     <div className="w-full">
       <div
-        className="wow fadeInUp rounded-sm bg-white p-8 shadow-two duration-300 hover:shadow-one dark:bg-dark dark:shadow-three dark:hover:shadow-gray-dark lg:px-5 xl:px-8"
+        className="wow fadeInUp shadow-two hover:shadow-one dark:bg-dark dark:shadow-three dark:hover:shadow-gray-dark rounded-sm bg-white p-8 duration-300 lg:px-5 xl:px-8"
         data-wow-delay=".1s"
       >
         <div className="mb-5 flex items-center space-x-1">{ratingIcons}</div>
-        <p className="mb-8 border-b border-body-color border-opacity-10 pb-8 text-base leading-relaxed text-body-color dark:border-white dark:border-opacity-10 dark:text-white">
+        <p className="border-body-color border-opacity-10 text-body-color dark:border-opacity-10 mb-8 border-b pb-8 text-base leading-relaxed dark:border-white dark:text-white">
           “{content}
         </p>
         <div className="flex items-center">
           <div className="relative mr-4 h-[50px] w-full max-w-[50px] overflow-hidden rounded-full">
-            <Image src={image} alt={name} fill />
+            <Image src={image} alt={name} fill sizes="50px" />
           </div>
           <div className="w-full">
-            <h3 className="mb-1 text-lg font-semibold text-dark dark:text-white lg:text-base xl:text-lg">{name}</h3>
-            <p className="text-sm text-body-color">{designation}</p>
+            <h3 className="text-dark mb-1 text-lg font-semibold lg:text-base xl:text-lg dark:text-white">{name}</h3>
+            <p className="text-body-color text-sm">{designation}</p>
           </div>
         </div>
       </div>
