@@ -1,6 +1,6 @@
 export function TopBanner() {
   return (
-    <div className="z-[100] hidden h-[40px] w-full flex-col items-center justify-center bg-gray-100 py-2 text-base dark:bg-gray-700 sm:flex-row sm:py-0 md:text-lg lg:flex">
+    <div className="z-[100] hidden h-[40px] w-full flex-col items-center justify-center bg-gray-100 py-2 text-base sm:flex-row sm:py-0 md:text-lg lg:flex dark:bg-gray-700">
       <div className="hidden sm:block">Support Ukraine &#127482;&#127462;</div>
       <a
         href="https://opensource.fb.com/support-ukraine"
