@@ -1,7 +1,7 @@
 export default function LoginForm() {
   return (
     <>
-      <button className="mb-6 flex w-full items-center justify-center rounded-sm border border-stroke bg-[#f8f8f8] px-6 py-3 text-base text-body-color outline-none transition-all duration-300 hover:border-primary hover:bg-primary/5 hover:text-primary dark:border-transparent dark:bg-[#2C303B] dark:text-body-color-dark dark:shadow-two dark:hover:border-primary dark:hover:bg-primary/5 dark:hover:text-primary dark:hover:shadow-none">
+      <button className="border-stroke text-body-color hover:border-primary hover:bg-primary/5 hover:text-primary dark:text-body-color-dark dark:shadow-two dark:hover:border-primary dark:hover:bg-primary/5 dark:hover:text-primary mb-6 flex w-full items-center justify-center rounded-sm border bg-[#f8f8f8] px-6 py-3 text-base transition-all duration-300 outline-none dark:border-transparent dark:bg-[#2C303B] dark:hover:shadow-none">
         <span className="mr-3">
           <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
             <g clipPath="url(#clip0_95:967)">
@@ -32,7 +32,7 @@ export default function LoginForm() {
         Sign in with Google
       </button>
 
-      <button className="mb-6 flex w-full items-center justify-center rounded-sm border border-stroke bg-[#f8f8f8] px-6 py-3 text-base text-body-color outline-none transition-all duration-300 hover:border-primary hover:bg-primary/5 hover:text-primary dark:border-transparent dark:bg-[#2C303B] dark:text-body-color-dark dark:shadow-two dark:hover:border-primary dark:hover:bg-primary/5 dark:hover:text-primary dark:hover:shadow-none">
+      <button className="border-stroke text-body-color hover:border-primary hover:bg-primary/5 hover:text-primary dark:text-body-color-dark dark:shadow-two dark:hover:border-primary dark:hover:bg-primary/5 dark:hover:text-primary mb-6 flex w-full items-center justify-center rounded-sm border bg-[#f8f8f8] px-6 py-3 text-base transition-all duration-300 outline-none dark:border-transparent dark:bg-[#2C303B] dark:hover:shadow-none">
         <span className="mr-3">
           <svg fill="currentColor" width="22" height="22" viewBox="0 0 64 64" xmlns="http://www.w3.org/2000/svg">
             <path d="M32 1.7998C15 1.7998 1 15.5998 1 32.7998C1 46.3998 9.9 57.9998 22.3 62.1998C23.9 62.4998 24.4 61.4998 24.4 60.7998C24.4 60.0998 24.4 58.0998 24.3 55.3998C15.7 57.3998 13.9 51.1998 13.9 51.1998C12.5 47.6998 10.4 46.6998 10.4 46.6998C7.6 44.6998 10.5 44.6998 10.5 44.6998C13.6 44.7998 15.3 47.8998 15.3 47.8998C18 52.6998 22.6 51.2998 24.3 50.3998C24.6 48.3998 25.4 46.9998 26.3 46.1998C19.5 45.4998 12.2 42.7998 12.2 30.9998C12.2 27.5998 13.5 24.8998 15.4 22.7998C15.1 22.0998 14 18.8998 15.7 14.5998C15.7 14.5998 18.4 13.7998 24.3 17.7998C26.8 17.0998 29.4 16.6998 32.1 16.6998C34.8 16.6998 37.5 16.9998 39.9 17.7998C45.8 13.8998 48.4 14.5998 48.4 14.5998C50.1 18.7998 49.1 22.0998 48.7 22.7998C50.7 24.8998 51.9 27.6998 51.9 30.9998C51.9 42.7998 44.6 45.4998 37.8 46.1998C38.9 47.1998 39.9 49.1998 39.9 51.9998C39.9 56.1998 39.8 59.4998 39.8 60.4998C39.8 61.2998 40.4 62.1998 41.9 61.8998C54.1 57.7998 63 46.2998 63 32.5998C62.9 15.5998 49 1.7998 32 1.7998Z" />
@@ -41,42 +41,42 @@ export default function LoginForm() {
         Sign in with Github
       </button>
       <div className="mb-8 flex items-center justify-center">
-        <span className="hidden h-[1px] w-full max-w-[70px] bg-body-color/50 sm:block"></span>
-        <p className="w-full px-5 text-center text-base font-medium text-body-color">Or, sign in with your email</p>
-        <span className="hidden h-[1px] w-full max-w-[70px] bg-body-color/50 sm:block"></span>
+        <span className="bg-body-color/50 hidden h-[1px] w-full max-w-[70px] sm:block"></span>
+        <p className="text-body-color w-full px-5 text-center text-base font-medium">Or, sign in with your email</p>
+        <span className="bg-body-color/50 hidden h-[1px] w-full max-w-[70px] sm:block"></span>
       </div>
       <form>
         <div className="mb-8">
-          <label htmlFor="email" className="mb-3 block text-sm text-dark dark:text-white">
+          <label htmlFor="email" className="text-dark mb-3 block text-sm dark:text-white">
             Your Email
           </label>
           <input
             type="email"
             name="email"
             placeholder="Enter your Email"
-            className="w-full rounded-sm border border-stroke bg-[#f8f8f8] px-6 py-3 text-base text-body-color outline-none transition-all duration-300 focus:border-primary dark:border-transparent dark:bg-[#2C303B] dark:text-body-color-dark dark:shadow-two dark:focus:border-primary dark:focus:shadow-none"
+            className="border-stroke text-body-color focus:border-primary dark:text-body-color-dark dark:shadow-two dark:focus:border-primary w-full rounded-sm border bg-[#f8f8f8] px-6 py-3 text-base transition-all duration-300 outline-none dark:border-transparent dark:bg-[#2C303B] dark:focus:shadow-none"
           />
         </div>
         <div className="mb-8">
-          <label htmlFor="password" className="mb-3 block text-sm text-dark dark:text-white">
+          <label htmlFor="password" className="text-dark mb-3 block text-sm dark:text-white">
             Your Password
           </label>
           <input
             type="password"
             name="password"
             placeholder="Enter your Password"
-            className="w-full rounded-sm border border-stroke bg-[#f8f8f8] px-6 py-3 text-base text-body-color outline-none transition-all duration-300 focus:border-primary dark:border-transparent dark:bg-[#2C303B] dark:text-body-color-dark dark:shadow-two dark:focus:border-primary dark:focus:shadow-none"
+            className="border-stroke text-body-color focus:border-primary dark:text-body-color-dark dark:shadow-two dark:focus:border-primary w-full rounded-sm border bg-[#f8f8f8] px-6 py-3 text-base transition-all duration-300 outline-none dark:border-transparent dark:bg-[#2C303B] dark:focus:shadow-none"
           />
         </div>
         <div className="mb-8 flex flex-col justify-between sm:flex-row sm:items-center">
           <div className="mb-4 sm:mb-0">
             <label
               htmlFor="checkboxLabel"
-              className="flex cursor-pointer select-none items-center text-sm font-medium text-body-color"
+              className="text-body-color flex cursor-pointer items-center text-sm font-medium select-none"
             >
               <div className="relative">
                 <input type="checkbox" id="checkboxLabel" className="sr-only" />
-                <div className="box mr-4 flex h-5 w-5 items-center justify-center rounded border border-body-color border-opacity-20 dark:border-white dark:border-opacity-10">
+                <div className="box border-body-color border-opacity-20 dark:border-opacity-10 mr-4 flex h-5 w-5 items-center justify-center rounded border dark:border-white">
                   <span className="opacity-0">
                     <svg width="11" height="8" viewBox="0 0 11 8" fill="none" xmlns="http://www.w3.org/2000/svg">
                       <path
@@ -93,13 +93,13 @@ export default function LoginForm() {
             </label>
           </div>
           <div>
-            <a href="#0" className="text-sm font-medium text-primary hover:underline">
+            <a href="#0" className="text-primary text-sm font-medium hover:underline">
               Forgot Password?
             </a>
           </div>
         </div>
         <div className="mb-6">
-          <button className="flex w-full items-center justify-center rounded-sm bg-primary px-9 py-4 text-base font-medium text-white shadow-submit duration-300 hover:bg-primary/90 dark:shadow-submit-dark">
+          <button className="bg-primary shadow-submit hover:bg-primary/90 dark:shadow-submit-dark flex w-full items-center justify-center rounded-sm px-9 py-4 text-base font-medium text-white duration-300">
             Sign in
           </button>
         </div>
