@@ -1,6 +1,8 @@
+import { ReactElement } from 'react';
+
 export type Feature = {
   id: number;
-  icon: JSX.Element;
+  icon: ReactElement;
   title: string;
   paragraph: string;
 };
